feat: Allow configuration of protected frontend routes
- Split out Page component from Router module
- Move ConnectivityWarning from Page to App component
- Add a boolean `protected` property to the route config objects
- Use the protected property from each route config to display the route component or
  redirect to the login route
- Add login and signup routes to frontend router config (WIP)
