+++ agir/front/components/app/App.js
@@ -5,10 +5,12 @@ import { GlobalContextProvider } from "@agir/front/globalContext/GlobalContext";
 import Router from "./Router";
 import TopBar from "@agir/front/allPages/TopBar";
 import PushModal from "@agir/front/allPages/PushModal";
+import ConnectivityWarning from "@agir/front/app/ConnectivityWarning";
 
 export default function App() {
   return (
     <GlobalContextProvider>
+      <ConnectivityWarning />
       <Router>
         <Helmet>
           <title>Plateforme d'action - Action populaire</title>
