+++ agir/front/components/authentication/SignupPage.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const SignupPage = () => {
+  return <div>Inscription !</div>;
+};
+
+export default SignupPage;
